IGNITE-14368 Data structures system views.
