+++ modules/core/src/main/java/org/apache/ignite/internal/managers/systemview/walker/AtomicLongViewWalker.java
@@ -17,7 +17,7 @@
 
 package org.apache.ignite.internal.managers.systemview.walker;
 
-import org.apache.ignite.spi.systemview.view.AtomicLongView;
+import org.apache.ignite.spi.systemview.view.datastructures.AtomicLongView;
 import org.apache.ignite.spi.systemview.view.SystemViewRowAttributeWalker;
 
 /**
