+++ modules/core/src/main/java/org/apache/ignite/internal/managers/systemview/walker/AtomicSequenceViewWalker.java
@@ -17,7 +17,7 @@
 
 package org.apache.ignite.internal.managers.systemview.walker;
 
-import org.apache.ignite.spi.systemview.view.AtomicSequenceView;
+import org.apache.ignite.spi.systemview.view.datastructures.AtomicSequenceView;
 import org.apache.ignite.spi.systemview.view.SystemViewRowAttributeWalker;
 
 /**
