+++ modules/core/src/main/java/org/apache/ignite/internal/managers/systemview/walker/AtomicStampedViewWalker.java
@@ -17,7 +17,7 @@
 
 package org.apache.ignite.internal.managers.systemview.walker;
 
-import org.apache.ignite.spi.systemview.view.AtomicStampedView;
+import org.apache.ignite.spi.systemview.view.datastructures.AtomicStampedView;
 import org.apache.ignite.spi.systemview.view.SystemViewRowAttributeWalker;
 
 /**
