+++ modules/core/src/main/java/org/apache/ignite/internal/managers/systemview/walker/CountDownLatchViewWalker.java
@@ -17,7 +17,7 @@
 
 package org.apache.ignite.internal.managers.systemview.walker;
 
-import org.apache.ignite.spi.systemview.view.CountDownLatchView;
+import org.apache.ignite.spi.systemview.view.datastructures.CountDownLatchView;
 import org.apache.ignite.spi.systemview.view.SystemViewRowAttributeWalker;
 
 /**
