+++ modules/core/src/main/java/org/apache/ignite/internal/managers/systemview/walker/QueueViewWalker.java
@@ -18,7 +18,7 @@
 package org.apache.ignite.internal.managers.systemview.walker;
 
 import org.apache.ignite.lang.IgniteUuid;
-import org.apache.ignite.spi.systemview.view.QueueView;
+import org.apache.ignite.spi.systemview.view.datastructures.QueueView;
 import org.apache.ignite.spi.systemview.view.SystemViewRowAttributeWalker;
 
 /**
