+++ modules/core/src/main/java/org/apache/ignite/internal/managers/systemview/walker/ReentrantLockViewWalker.java
@@ -17,7 +17,7 @@
 
 package org.apache.ignite.internal.managers.systemview.walker;
 
-import org.apache.ignite.spi.systemview.view.ReentrantLockView;
+import org.apache.ignite.spi.systemview.view.datastructures.ReentrantLockView;
 import org.apache.ignite.spi.systemview.view.SystemViewRowAttributeWalker;
 
 /**
