+++ modules/core/src/main/java/org/apache/ignite/internal/managers/systemview/walker/SetViewWalker.java
@@ -18,7 +18,7 @@
 package org.apache.ignite.internal.managers.systemview.walker;
 
 import org.apache.ignite.lang.IgniteUuid;
-import org.apache.ignite.spi.systemview.view.SetView;
+import org.apache.ignite.spi.systemview.view.datastructures.SetView;
 import org.apache.ignite.spi.systemview.view.SystemViewRowAttributeWalker;
 
 /**
