+++ modules/core/src/main/java/org/apache/ignite/spi/systemview/view/datastructures/AbstractDataStructureView.java
@@ -15,11 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.spi.systemview.view;
+package org.apache.ignite.spi.systemview.view.datastructures;
 
 import org.apache.ignite.internal.managers.systemview.walker.Order;
 import org.apache.ignite.internal.processors.datastructures.AtomicDataStructureProxy;
 import org.apache.ignite.internal.util.typedef.internal.CU;
+import org.apache.ignite.spi.systemview.view.SystemView;
 
 /** Abstract class for a {@link SystemView} representation of data structures. */
 abstract class AbstractDataStructureView<T extends AtomicDataStructureProxy> {
