+++ modules/core/src/main/java/org/apache/ignite/spi/systemview/view/datastructures/AtomicReferenceView.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.spi.systemview.view;
+package org.apache.ignite.spi.systemview.view.datastructures;
 
 import org.apache.ignite.Ignite;
 import org.apache.ignite.IgniteAtomicReference;
@@ -23,6 +23,7 @@ import org.apache.ignite.configuration.AtomicConfiguration;
 import org.apache.ignite.internal.managers.systemview.walker.Order;
 import org.apache.ignite.internal.processors.datastructures.GridCacheAtomicReferenceImpl;
 import org.apache.ignite.internal.processors.datastructures.GridCacheRemovable;
+import org.apache.ignite.spi.systemview.view.SystemView;
 
 import static org.apache.ignite.internal.util.IgniteUtils.toStringSafe;
