+++ modules/core/src/main/java/org/apache/ignite/spi/systemview/view/datastructures/AtomicStampedView.java
@@ -15,42 +15,45 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.spi.systemview.view;
+package org.apache.ignite.spi.systemview.view.datastructures;
 
 import org.apache.ignite.Ignite;
-import org.apache.ignite.IgniteAtomicSequence;
+import org.apache.ignite.IgniteAtomicStamped;
 import org.apache.ignite.configuration.AtomicConfiguration;
 import org.apache.ignite.internal.managers.systemview.walker.Order;
-import org.apache.ignite.internal.processors.datastructures.GridCacheAtomicSequenceImpl;
+import org.apache.ignite.internal.processors.datastructures.GridCacheAtomicStampedImpl;
 import org.apache.ignite.internal.processors.datastructures.GridCacheRemovable;
+import org.apache.ignite.spi.systemview.view.SystemView;
+
+import static org.apache.ignite.internal.util.IgniteUtils.toStringSafe;
 
 /**
- * {@link IgniteAtomicSequence} representation for a {@link SystemView}.
+ * {@link IgniteAtomicStamped} representation for a {@link SystemView}.
  *
- * @see Ignite#atomicSequence(String, long, boolean)
- * @see Ignite#atomicSequence(String, AtomicConfiguration, long, boolean)
+ * @see Ignite#atomicStamped(String, Object, Object, boolean)
+ * @see Ignite#atomicStamped(String, AtomicConfiguration, Object, Object, boolean)
  */
-public class AtomicSequenceView extends AbstractDataStructureView<GridCacheAtomicSequenceImpl> {
+public class AtomicStampedView extends AbstractDataStructureView<GridCacheAtomicStampedImpl> {
     /** @param ds Data structure instance. */
-    public AtomicSequenceView(GridCacheRemovable ds) {
-        super((GridCacheAtomicSequenceImpl)ds);
+    public AtomicStampedView(GridCacheRemovable ds) {
+        super((GridCacheAtomicStampedImpl)ds);
     }
 
     /**
      * @return Value.
-     * @see IgniteAtomicSequence#get()
+     * @see IgniteAtomicStamped#get()
      */
     @Order(1)
-    public long value() {
-        return ds.get();
+    public String value() {
+        return toStringSafe(ds.value());
     }
 
     /**
-     * @return Batch size.
-     * @see IgniteAtomicSequence#batchSize()
+     * @return Stamp.
+     * @see IgniteAtomicStamped#stamp()
      */
     @Order(2)
-    public long batchSize() {
-        return ds.batchSize();
+    public String stamp() {
+        return toStringSafe(ds.stamp());
     }
 }
