+++ modules/core/src/main/java/org/apache/ignite/spi/systemview/view/datastructures/QueueView.java
@@ -15,11 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.spi.systemview.view;
+package org.apache.ignite.spi.systemview.view.datastructures;
 
 import org.apache.ignite.internal.managers.systemview.walker.Order;
 import org.apache.ignite.internal.processors.datastructures.GridCacheQueueProxy;
 import org.apache.ignite.lang.IgniteUuid;
+import org.apache.ignite.spi.systemview.view.SystemView;
 
 /**
  * Queue representation for a {@link SystemView}.
