+++ modules/core/src/main/java/org/apache/ignite/spi/systemview/view/datastructures/SemaphoreView.java
@@ -15,13 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.spi.systemview.view;
+package org.apache.ignite.spi.systemview.view.datastructures;
 
 import org.apache.ignite.Ignite;
 import org.apache.ignite.IgniteSemaphore;
 import org.apache.ignite.internal.managers.systemview.walker.Order;
 import org.apache.ignite.internal.processors.datastructures.GridCacheRemovable;
 import org.apache.ignite.internal.processors.datastructures.GridCacheSemaphoreImpl;
+import org.apache.ignite.spi.systemview.view.SystemView;
 
 /**
  * {@link IgniteSemaphore} representation for a {@link SystemView}.
